+++ install.ps1
@@ -2,4 +2,4 @@
 Param ()
 
 Install-Module -Name PSDepend -Force
-Invoke-PSDepend -Force -Verbose
+Invoke-PSDepend -Force